+++ NetworkOnline/GameScene.go
@@ -70,8 +70,8 @@ func (s *GameScene) Load() {
 	FPSDrawer := Engine.NewGameObject("FPS")
 	txt := FPSDrawer.AddComponent(Components.NewUIText(ArialFont2, "")).(*Components.UIText)
 	fps := FPSDrawer.AddComponent(Engine.NewFPS()).(*Engine.FPS)
-	fps.SetAction(func(fps float32) {
-		txt.SetString("FPS: " + strconv.FormatFloat(float64(fps), 'f', 2, 32))
+	fps.SetAction(func(fps float64) {
+		txt.SetString("FPS: " + strconv.FormatFloat(fps, 'f', 2, 32))
 	})
 	FPSDrawer.Transform().SetParent2(cam)
 	FPSDrawer.Transform().SetPositionf(60, float32(Engine.Height)-20)
